Added abstract OMPLPlanningContext + have ModelBasedPlanningContext inherit from OMPLPlanningContext
